requirements.txt
streamlit
langchain
google-genai
langchain-google-genai
unstructured
python-dotenv
faiss-cpu
